+++ src/components/User.tsx
@@ -1,4 +1,6 @@
-import { sessionApi } from '../components/sessionApi';
+import { sessionApiLogin } from './sessionApiLogin';
+import { sessionApiLogout } from './sessionApiLogout';
+
 // import { useState } from 'react';
 import Cookies from 'js-cookie';
 
@@ -8,20 +10,12 @@ class User {
   set = (key: string, value: string) => Cookies.set(key, value);
   get = (key: string) => Cookies.get(key);
 
-  // getLocalStorage = (key: string) => {
-  //   const ret = localStorage.getItem(key);
-  //   if (ret) {
-  //     return ret;
-  //   }
-  //   return null;
-  // };
-
   login = async (email: string, password: string) => {
     if (process.env.NODE_ENV !== 'production') {
       console.log(email);
       console.log(password);
     }
-    await sessionApi.login({ email, password });
+    await sessionApiLogin({ email, password });
     if (process.env.NODE_ENV !== 'production') {
       console.log('isLoggedIn(User.tsxログイン処理後):' + this.isLoggedIn());
       console.log('this.get(User.tsx):', this.get('isLoggedIn'));
@@ -30,14 +24,9 @@ class User {
   };
 
   logout = async () => {
-    console.log('User.logout ');
-
     if (this.isLoggedIn()) {
       this.set('isLoggedIn', false.toString());
-      sessionApi.logout;
-
-      // ログアウト処理
-      // 他に必要な処理があるのならこちら
+      sessionApiLogout;
     }
   };
 }
